refactor(operatorgroup): split apart the larger sync functions
also fixed an issue where all deployments in a namespace were being
annotated, not just those owned by CSVs
